pyproject.toml
[project]
name = "nbt-image-classification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ollama>=0.4.0",
    "pillow>=10.0.0",
]
